requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Dance==5.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1